LayoutTests/http/tests/security/video-cross-origin-accessfailure-expected.txt

CONSOLE MESSAGE: Cross-origin media resource load denied by Cross-Origin Resource Sharing policy.
EVENT(error)
END OF TEST
 
